requirements.txt
asgiref==3.8.1
Django==5.1.2
django-filter==25.1
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.16.1
lxml==5.3.0
Markdown==3.8.2
numpy==2.1.2
pandas==2.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
aiogram==3.4.1
aiohttp==3.9.5
cryptg==0.4.0
python-dotenv==1.0.1